Move item store component to separate file
